vps-audit/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
paramiko==3.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.1
cryptography==41.0.5
python-multipart==0.0.6
requests==2.31.0
asyncpg==0.29.0
aiohttp==3.9.1
tenacity==8.2.3
alembic==1.12.1
greenlet==3.0.1
httpx==0.25.2